refactor(vector_stores): extract ChromaDBLocalVectorStore
Move chunking, embedding, chroma client management and the ingest /
search / delete / healthcheck path out of LocalFSChromaDBDatasetType
and into a new ChromaDBLocalVectorStore in components/vector_stores/
chromadb_local/.

The dataset type now composes a LocalFileSource + a
ChromaDBLocalVectorStore and delegates lifecycle methods. The combined
ChromaDBLocalConfiguration stays at the binding layer; the vector
store has its own narrow ChromaDBLocalVectorStoreConfiguration
covering only collection_name + http_port.

Provisioner stays in dataset_types/chromadb_local/ for now —
re-keying provisioner registration is deferred to the cleanup PR to
avoid coupling concerns.
