+++ backend/syft_space/components/vector_stores/chromadb_local/__init__.py
@@ -1,0 +1,13 @@
+"""ChromaDB local vector store package."""
+
+from syft_space.components.vector_stores.chromadb_local.chromadb_vector_store import (
+    ChromaDBLocalVectorStore,
+)
+from syft_space.components.vector_stores.chromadb_local.schemas import (
+    ChromaDBLocalVectorStoreConfiguration,
+)
+
+__all__ = [
+    "ChromaDBLocalVectorStore",
+    "ChromaDBLocalVectorStoreConfiguration",
+]
